+++ src-tauri/src/database/repositories/task_repo.rs
@@ -1,22 +1,20 @@
-use crate::models::database::filters::TaskFilters;
+use crate::{
+    database::repositories::task_builder::TaskQueryBuilder,
+    models::database::filters::TaskFilters,
+};
 use crate::models::database::stats::TaskStats;
 use crate::models::database::task::Task;
-use sqlx::{Pool, Sqlite};
+use sqlx::{ Pool, Sqlite };
 
-pub async fn get_stats(pool: &Pool<Sqlite>, category: &str) -> Result<TaskStats, sqlx::Error> {
-    let row = sqlx::query_as::<_, (i32, i32)>(
-        r#"
-        SELECT 
-            COUNT(*), 
-            CAST(COALESCE(SUM(current_completions >= max_completions), 0) AS INTEGER) 
-        FROM task_tracker 
-        WHERE (category = ? OR ? = 'All')
-        "#,
-    )
-    .bind(category)
-    .bind(category)
-    .fetch_one(pool)
-    .await?;
+pub async fn get_stats(
+    pool: &Pool<Sqlite>,
+    category: &str,
+    filters: &TaskFilters
+) -> Result<TaskStats, sqlx::Error> {
+    let row: (i32, i32) = TaskQueryBuilder::new(category, "", filters)
+        .build_stats()
+        .build_query_as::<(i32, i32)>()
+        .fetch_one(pool).await?;
 
     Ok(TaskStats {
         current: row.1,
@@ -30,59 +28,32 @@ pub async fn find_all(
     search: &str,
     filters: &TaskFilters,
     limit: i64,
-    offset: i64,
+    offset: i64
 ) -> Result<Vec<Task>, sqlx::Error> {
-    let search_pattern = format!("%{}%", search);
-
-    sqlx::query_as::<_, Task>(
-        r#"
-        SELECT * FROM task_tracker 
-        WHERE (category = ? OR ? = 'All')
-        AND (name LIKE ? OR tags LIKE ? OR location LIKE ?)
-        AND (NOT (? AND current_completions < max_completions))
-        AND (NOT (? AND current_completions >= max_completions))
-        AND (NOT (? AND favorite = 1))
-        AND (NOT (? AND favorite = 0))
-        ORDER BY 
-            -- Sort by favorite if the filter is active, then by name
-            CASE WHEN ? THEN favorite END DESC, 
-            name ASC
-        LIMIT ? OFFSET ?
-        "#,
-    )
-    .bind(category)
-    .bind(category)
-    .bind(&search_pattern)
-    .bind(&search_pattern)
-    .bind(&search_pattern)
-    .bind(filters.hide_incomplete)
-    .bind(filters.hide_complete)
-    .bind(filters.hide_favorite)
-    .bind(filters.hide_non_favorite)
-    .bind(filters.favorite_first)
-    .bind(limit)
-    .bind(offset)
-    .fetch_all(pool)
-    .await
+    TaskQueryBuilder::new(category, search, filters)
+        .paginate(limit, offset)
+        .build()
+        .build_query_as::<Task>()
+        .fetch_all(pool).await
 }
 
 pub async fn find_by_id(pool: &Pool<Sqlite>, id: &str) -> Result<Task, sqlx::Error> {
-    sqlx::query_as::<_, Task>("SELECT * FROM task_tracker WHERE id = ?")
+    sqlx
+        ::query_as::<_, Task>("SELECT * FROM task_tracker WHERE id = ?")
         .bind(id)
-        .fetch_one(pool)
-        .await
+        .fetch_one(pool).await
 }
 
 pub async fn set_favorite_status(
     pool: &Pool<Sqlite>,
     id: &str,
-    is_favorite: bool,
+    is_favorite: bool
 ) -> Result<(), sqlx::Error> {
-    sqlx::query("UPDATE task_tracker SET favorite = ? WHERE id = ?")
+    sqlx
+        ::query("UPDATE task_tracker SET favorite = ? WHERE id = ?")
         .bind(is_favorite)
         .bind(id)
-        .execute(pool)
-        .await?;
+        .execute(pool).await?;
 
     Ok(())
 }
@@ -91,10 +62,11 @@ pub async fn update_completions(
     pool: &Pool<Sqlite>,
     id: &str,
     count: i32,
-    reset_time: String,
+    reset_time: String
 ) -> Result<Task, sqlx::Error> {
-    sqlx::query_as::<_, Task>(
-        r#"
+    sqlx
+        ::query_as::<_, Task>(
+            r#"
         UPDATE task_tracker 
         SET current_completions = CASE 
             WHEN ? > max_completions THEN max_completions 
@@ -104,34 +76,30 @@ pub async fn update_completions(
         last_reset = ?
         WHERE id = ?
         RETURNING *
-        "#,
-    )
-    .bind(count)
-    .bind(count)
-    .bind(count)
-    .bind(reset_time)
-    .bind(id)
-    .fetch_one(pool)
-    .await
+        "#
+        )
+        .bind(count)
+        .bind(count)
+        .bind(count)
+        .bind(reset_time)
+        .bind(id)
+        .fetch_one(pool).await
 }
 
 pub async fn find_all_raw(pool: &Pool<Sqlite>) -> Result<Vec<Task>, sqlx::Error> {
-    sqlx::query_as::<_, Task>("SELECT * FROM task_tracker")
-        .fetch_all(pool)
-        .await
+    sqlx::query_as::<_, Task>("SELECT * FROM task_tracker").fetch_all(pool).await
 }
 
 pub async fn reset_task_progress(
     pool: &Pool<Sqlite>,
     id: &str,
-    reset_time: String,
+    reset_time: String
 ) -> Result<u64, sqlx::Error> {
-    let res =
-        sqlx::query("UPDATE task_tracker SET current_completions = 0, last_reset = ? WHERE id = ?")
-            .bind(reset_time)
-            .bind(id)
-            .execute(pool)
-            .await?;
+    let res = sqlx
+        ::query("UPDATE task_tracker SET current_completions = 0, last_reset = ? WHERE id = ?")
+        .bind(reset_time)
+        .bind(id)
+        .execute(pool).await?;
 
     Ok(res.rows_affected())
 }