+++ CHANGELOG.md
@@ -9,6 +9,10 @@ All notable changes to this project will be logged here
 - 🛠️ fix Mastery items not always showing the right amount of the component required
 - 🛠️ fix searchpulse math
 
+### Dev Notes
+
+- ✏️ move to query builder for some repo functions in `mastery_repo.rs` and `task_repo.rs`
+
 ## 1.4.4
 
 ### Features